requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
django-rosetta==0.9.9
djangorestframework==3.14.0
environs==9.5.0
idna==3.4
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.1
phonenumberslite==8.13.16
Pillow==10.0.0
polib==1.2.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
